requirements.txt
# 依赖清单
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
python-docx>=0.8.11
openpyxl>=3.1.2
pandas>=2.1.0
faiss-cpu>=1.7.4
langchain>=0.1.0